// frontend/src/app/api/model/chatMessageDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RagSourceDto } from './ragSourceDto';


export interface ChatMessageDto { 
    id?: string;
    role?: string | null;
    content?: string | null;
    createdAtUtc?: string;
    sources?: Array<RagSourceDto> | null;
}