// frontend/src/app/api/model/threadSummaryDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ThreadSummaryDto { 
    id?: string;
    title?: string | null;
    updatedAtUtc?: string;
}